# Conflicts:
#   src/router/index.js
#   src/views/emojiPacks/index.vue
#   src/views/invites/index.vue
#   src/views/layout/components/Sidebar/Item.vue

+++ src/views/users/components/ModerationDropdown.vue
@@ -38,6 +38,17 @@
         {{ $t('users.deleteAccount') }}
       </el-dropdown-item>
       <el-dropdown-item
+        v-if="user.local && user.approval_pending"
+        divided
+        @click.native="handleAccountApproval(user)">
+        {{ $t('users.approveAccount') }}
+      </el-dropdown-item>
+      <el-dropdown-item
+        v-if="user.local && user.approval_pending"
+        @click.native="handleAccountRejection(user)">
+        {{ $t('users.rejectAccount') }}
+      </el-dropdown-item>
+      <el-dropdown-item
         v-if="user.local && user.confirmation_pending"
         divided
         @click.native="handleEmailConfirmation(user)">
@@ -50,42 +61,42 @@
       </el-dropdown-item>
       <el-dropdown-item
         :divided="showAdminAction(user)"
-        :class="{ 'active-tag': user.tags.includes('force_nsfw') }"
-        @click.native="toggleTag(user, 'force_nsfw')">
+        :class="{ 'active-tag': user.tags.includes('mrf_tag:media-force-nsfw') }"
+        @click.native="toggleTag(user, 'mrf_tag:media-force-nsfw')">
         {{ $t('users.forceNsfw') }}
-        <i v-if="user.tags.includes('force_nsfw')" class="el-icon-check"/>
+        <i v-if="user.tags.includes('mrf_tag:media-force-nsfw')" class="el-icon-check"/>
       </el-dropdown-item>
       <el-dropdown-item
-        :class="{ 'active-tag': user.tags.includes('strip_media') }"
-        @click.native="toggleTag(user, 'strip_media')">
+        :class="{ 'active-tag': user.tags.includes('mrf_tag:media-strip') }"
+        @click.native="toggleTag(user, 'mrf_tag:media-strip')">
         {{ $t('users.stripMedia') }}
-        <i v-if="user.tags.includes('strip_media')" class="el-icon-check"/>
+        <i v-if="user.tags.includes('mrf_tag:media-strip')" class="el-icon-check"/>
       </el-dropdown-item>
       <el-dropdown-item
-        :class="{ 'active-tag': user.tags.includes('force_unlisted') }"
-        @click.native="toggleTag(user, 'force_unlisted')">
+        :class="{ 'active-tag': user.tags.includes('mrf_tag:force-unlisted') }"
+        @click.native="toggleTag(user, 'mrf_tag:force-unlisted')">
         {{ $t('users.forceUnlisted') }}
-        <i v-if="user.tags.includes('force_unlisted')" class="el-icon-check"/>
+        <i v-if="user.tags.includes('mrf_tag:force-unlisted')" class="el-icon-check"/>
       </el-dropdown-item>
       <el-dropdown-item
-        :class="{ 'active-tag': user.tags.includes('sandbox') }"
-        @click.native="toggleTag(user, 'sandbox')">
+        :class="{ 'active-tag': user.tags.includes('mrf_tag:sandbox') }"
+        @click.native="toggleTag(user, 'mrf_tag:sandbox')">
         {{ $t('users.sandbox') }}
-        <i v-if="user.tags.includes('sandbox')" class="el-icon-check"/>
+        <i v-if="user.tags.includes('mrf_tag:sandbox')" class="el-icon-check"/>
       </el-dropdown-item>
       <el-dropdown-item
         v-if="user.local"
-        :class="{ 'active-tag': user.tags.includes('disable_remote_subscription') }"
-        @click.native="toggleTag(user, 'disable_remote_subscription')">
+        :class="{ 'active-tag': user.tags.includes('mrf_tag:disable-remote-subscription') }"
+        @click.native="toggleTag(user, 'mrf_tag:disable-remote-subscription')">
         {{ $t('users.disableRemoteSubscription') }}
-        <i v-if="user.tags.includes('disable_remote_subscription')" class="el-icon-check"/>
+        <i v-if="user.tags.includes('mrf_tag:disable-remote-subscription')" class="el-icon-check"/>
       </el-dropdown-item>
       <el-dropdown-item
         v-if="user.local"
-        :class="{ 'active-tag': user.tags.includes('disable_any_subscription') }"
-        @click.native="toggleTag(user, 'disable_any_subscription')">
+        :class="{ 'active-tag': user.tags.includes('mrf_tag:disable-any-subscription') }"
+        @click.native="toggleTag(user, 'mrf_tag:disable-any-subscription')">
         {{ $t('users.disableAnySubscription') }}
-        <i v-if="user.tags.includes('disable_any_subscription')" class="el-icon-check"/>
+        <i v-if="user.tags.includes('mrf_tag:disable-any-subscription')" class="el-icon-check"/>
       </el-dropdown-item>
       <el-dropdown-item
         v-if="user.local"
@@ -97,6 +108,11 @@
         v-if="user.local"
         @click.native="requirePasswordReset(user)">
         {{ $t('users.requirePasswordReset') }}
+      </el-dropdown-item>
+      <el-dropdown-item
+        v-if="user.local"
+        @click.native="disableMfa(user.nickname)">
+        {{ $t('users.disableMfa') }}
       </el-dropdown-item>
     </el-dropdown-menu>
   </el-dropdown>
@@ -127,6 +143,9 @@ export default {
     }
   },
   methods: {
+    disableMfa(nickname) {
+      this.$store.dispatch('DisableMfa', nickname)
+    },
     getPasswordResetToken(nickname) {
       this.$emit('open-reset-token-dialog')
       this.$store.dispatch('GetPasswordResetToken', nickname)
@@ -135,7 +154,39 @@ export default {
       this.$store.dispatch('ResendConfirmationEmail', [user])
     },
     handleDeletion(user) {
-      this.$store.dispatch('DeleteUsers', { users: [user], _userId: user.id })
+      this.$confirm(
+        this.$t('users.deleteUserConfirmation'),
+        {
+          confirmButtonText: 'Delete',
+          cancelButtonText: 'Cancel',
+          type: 'warning'
+        }).then(() => {
+        this.$store.dispatch('DeleteUsers', { users: [user], _userId: user.id })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: 'Delete canceled'
+        })
+      })
+    },
+    handleAccountApproval(user) {
+      this.$store.dispatch('ApproveUsersAccount', { users: [user], _userId: user.id, _statusId: this.statusId })
+    },
+    handleAccountRejection(user) {
+      this.$confirm(
+        this.$t('users.rejectAccountConfirmation'),
+        {
+          confirmButtonText: 'Reject',
+          cancelButtonText: 'Cancel',
+          type: 'warning'
+        }).then(() => {
+        this.$store.dispatch('DeleteUsers', { users: [user], _userId: user.id })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: 'Reject canceled'
+        })
+      })
     },
     handleEmailConfirmation(user) {
       this.$store.dispatch('ConfirmUsersEmail', { users: [user], _userId: user.id, _statusId: this.statusId })
